backend/requirements.txt
fastapi==0.115.14
uvicorn==0.35.0
python-multipart==0.0.22
faster-whisper==1.1.1
librosa==0.11.0
soundfile==0.13.1
numpy==1.26.4
youtube-transcript-api==1.2.4
gemini_webapi==1.21.0
google-genai==1.28.0
aiohttp==3.12.15
sqlalchemy==2.0.23
slowapi==0.1.9
python-jose[cryptography]==3.3.0
python-dotenv==1.0.0
pytest==8.3.5
pytest-asyncio==0.24.0
httpx==0.28.1
